initial commit
by convention the root of the trunk MUST be titled "initial commit".

The keywords for MUST, MUST NOT, SHOULD, SHOULD NOT, MAY are in this and
future commits based on RFC 2119.

The above is a joke, but now thinking about it, maybe laying out git
commit message conventions in the "initial commit" is not such a bad
idea. Lets try:

The subject of the commit MUST be written in imperative present tense,
so that when merging commits to your local branch (eg. main <-
origin/main), it describes the changes the you would apply. It SHOULD be
no longer than 50 characters.
Exceptions to this include the "initial commit" or automated messages
through tooling, eg. when merging changes through the GitHub UI from
dependabot.

The body of the commit, SHOULD contain a motivation for the changes and
MAY also include technical explanation of otherwise obscure, but
sensible or idiomatic, changes. It SHOULD be longer than 80 characters.