create spike ns through bootstrap script
so that the test environment can consistently be setup across machines,
should include the setup of minikube eventually.

For now setup minikube with default settings and the ingress addon
enabled so that the registry can be tested with curl from the outside.

The reason for using a separate namespace "spike" is for easy clean up
and for testing persitence down the line.