test that the catalog is empty
since the sandbox for the test should be without any previous state, as
the teardown should include all the objects created for the test,
including the PV (which hopefully should have become better supported
in minikube by now).

Note that the default pathType for the ingress created with kubectl is
set to 'Exact'. I do not want to create individual rules for each path
in the ingress, so I set it to Prefix.

A potential downside of that solution is that any path for the registry
is reachable outside of the cluster.