transfer images from local registry to minikube registry
this might seem like an odd way of populating the registry we have
deployed to minikube with images for testing, but there did not seem to
be an easy way of doing this that was also somewhat portable.
Otherwise it would be required to change the local dns lookup to use
the minikube dns, but that would require changes in the local machine
setup, which is annoying and also not necessarily straightforward if
that machine is a GitHub runner.

In order to make it possible to transfer images from one registry to
another, digest by digest, blob by blob and manifest, the following flow
has to be followed:

GET manifest from source registry
-> extract the digest from the layers of the manifest
-> GET blobs by digest from source registry
-> POST upload to destination registry
-> extract location from response header
-> PUT blobs to destination registry using the location
-> PUT manifest to destination registry
The image has now been transferred from the source to the destination
registry.
See:
1. https://docs.docker.com/registry/spec/api/#get-manifest
2. https://docs.docker.com/registry/spec/api/#pulling-an-image
3. https://docs.docker.com/registry/spec/api/#content-digests
4. https://docs.docker.com/registry/spec/api/#get-blob
5. https://docs.docker.com/registry/spec/api/#initiate-blob-upload
6. https://docs.docker.com/registry/spec/api/#put-blob-upload
7. https://docs.docker.com/registry/spec/api/#put-manifest
bonus: https://docs.docker.com/registry/spec/manifest-v2-2/#media-types

Right now the above is handled by one function, which was put together
step by step without tests, not ideal, and likely not very robust, but
luckily interacting with the docker registry API is idempotent.
Nevertheless it might make sense to port it to Golang or some other
fitting language at some point to tests it and make it more robust.

Note that annotations was added to the registry ingress to fix the
error: "413 Entity too large" from the nginx ingress controller when
putting large blobs to the registry.
https://stackoverflow.com/questions/64759983/nginx-ingress-returns-413-entity-too-large