persist uploaded images in a PV
A bit surprised here that minikube creates a PV on the fly which
immediately gets bound to match the PVC.
So for now I am not creating a PV myself as it does not seem necessary.

Regarding the mountPath for the registry pods, the directory which
stores all the registry files (manifests, layers, uploads and blobs), is
configurable, but defaults to /var/lib/registry.
https://minikube.sigs.k8s.io/docs/handbook/persistent_volumes/
https://docs.docker.com/registry/storage-drivers/filesystem/