mount config from secret
So that the config can be modified to use redis for caching, which
hopefully should make it trivial to run multiple pods for the registry
without any concurrency issues.

config extracted from registry pod and generated with a kustomize
secretsGenerator.
```
$ kubectl -n spike exec -t registry-857968997c-8qjrx \
  -- cat /etc/docker/registry/config.yml > templates/config.yml
```
Which now means that the objects are created with the kubectl using its
built-in support for kustomize with `kustomize apply -k`
https://kubernetes.io/docs/tasks/configmap-secret/managing-secret-using-kustomize/