use redis for caching
I know it is working, because I have inspected the logs of the registry
in debug mode, but how to test it? Right now I have no idea, other than
accessing the logs. Maybe some performance tests are more relevant as
better performance seems to be the desired outcome, which makes sense...
I guess it would be possible to ask the redis API, but still we are not
interested in testing redis, rather in the outcome it has to use it for
caching.

It seems that using redis for a cache does nothing for any concurrency
issues, but rather "only" helps with the performance of accessing layer
metadata.
https://docs.docker.com/registry/configuration/#cache

The concurrency issues seems more likely related to a misconfiguration,
from the logs:
```
msg="No HTTP secret provided - generated random secret. This may cause
problems with uploads if multiple registries are behind a load-balancer.
To provide a shared secret, fill in http.secret in the configuration
file or set the REGISTRY_HTTP_SECRET environment variable."
```
At least now I know how to approach that issue, but still... Goddag mand
økseskaft.

https://redis.io/docs/manual/client-side-caching/
https://docs.redis.com/latest/rc/api/get-started/use-rest-api/
https://docs.docker.com/registry/configuration/#redis